feat: use current path and default name of 'output' for the output folder (#40)
